+++ Source/cielim/Actors/CaptureManager.cpp
