+++ Source/cielim/Actors/CelestialBody.cpp
