+++ Source/cielim/Actors/CelestialBody.h
