+++ Source/cielim/Scenes/SceneManager.h
@@ -12,8 +12,8 @@
 #include "CoreMinimal.h"
 
 #include "SceneData.h"
-#include "ZmqConnection/CielimCircularQueue.h"
-#include "ZmqConnection/QueueBridge.h"
+#include "../Network/CielimCircularQueue.h"
+#include "../Network/QueueBridge.h"
 
 #include "SceneManager.generated.h"
