+++ Source/cielim/Utilities/Logging/CielimLoggingMacros.cpp
