+++ Source/cielim/Utilities/Logging/CielimLoggingMacros.h
