+++ Source/cielim/Utilities/Math/KinematicsUtilities.cpp
